+++ src/platform/test/functional/fixtures/es_archiver/dashboard_elements/controls/chaining/data.json
@@ -1,0 +1,24 @@
+{
+  "type": "doc",
+  "value": {
+    "id": "control-chaining-cat",
+    "index": "animals-cats-2018-01-01",
+    "source": {
+      "@timestamp": "2018-01-01T16:00:00.000Z",
+      "animal": "cat"
+    }
+  }
+}
+
+{
+  "type": "doc",
+  "value": {
+    "id": "control-chaining-dog",
+    "index": "animals-dogs-2018-01-01",
+    "source": {
+      "@timestamp": "2018-01-01T16:00:00.000Z",
+      "name": "Max",
+      "sound": "woof"
+    }
+  }
+}
